Extract `debouncedGuess` to its own file
